move wallet reducer to own file
